test: move signing compatibility tests into the integration spec
The reference-selection and detached-signature cases exercise signing and
verification end to end, so they belong with the other integration tests
rather than in a spec of their own.
